Issue #11532 - Refactor test function, remove duplication
