+++ profile/modules/cp/modules/cp_taxonomy/tests/src/ExistingSiteJavascript/ApplyTermsNodeTest.php
@@ -3,6 +3,7 @@
 namespace Drupal\Tests\cp_taxonomy\ExistingSiteJavascript;
 
 use Drupal\node\Entity\Node;
+use Drupal\Tests\cp_taxonomy\Traits\CpTaxonomyTestTrait;
 
 /**
  * Tests taxonomy terms apply to nodes.
@@ -12,6 +13,8 @@ use Drupal\node\Entity\Node;
  * @covers \Drupal\cp_taxonomy\Form\AddTermsToNodeForm
  */
 class ApplyTermsNodeTest extends CpTaxonomyExistingSiteJavascriptTestBase {
+
+  use CpTaxonomyTestTrait;
 
   protected $term;
   protected $groupAdmin;
@@ -93,36 +96,6 @@ class ApplyTermsNodeTest extends CpTaxonomyExistingSiteJavascriptTestBase {
     $this->applyAction('cp_taxonomy_add_terms_node_action');
     $this->applyVocabularyFirstTerm($media_vocab);
     $web_assert->pageTextContains('Selected vocabulary is not handle node entity type.');
-  }
-
-  /**
-   * Helper function, that will apply the action.
-   */
-  protected function applyAction($action_id) {
-    $web_assert = $this->assertSession();
-    $page = $this->getCurrentPage();
-    $select = $page->findField('action');
-    $select->setValue($action_id);
-    $page->pressButton('Apply to selected items');
-    $web_assert->statusCodeEquals(200);
-  }
-
-  /**
-   * Helper function, that will select a vocab and first term in chosen.
-   */
-  protected function applyVocabularyFirstTerm($vocabulary) {
-    $web_assert = $this->assertSession();
-    $page = $this->getCurrentPage();
-    $select = $page->findField('vocabulary');
-    $select->setValue($vocabulary);
-    $this->waitForAjaxToFinish();
-    $page->find('css', '.chosen-search-input')->click();
-    $result = $web_assert->waitForElementVisible('css', '.active-result.highlighted');
-    $this->assertNotEmpty($result, 'Chosen popup is not visible.');
-    $page->find('css', '.active-result.highlighted')->click();
-    $page->find('css', '.chosen-search-input')->click();
-    $page->pressButton('Apply');
-    $web_assert->statusCodeEquals(200);
   }
 
 }
