+++ profile/modules/cp/modules/cp_taxonomy/tests/src/ExistingSiteJavascript/RemoveTermsPublicationTest.php
@@ -3,6 +3,7 @@
 namespace Drupal\Tests\cp_taxonomy\ExistingSiteJavascript;
 
 use Drupal\bibcite_entity\Entity\Reference;
+use Drupal\Tests\cp_taxonomy\Traits\CpTaxonomyTestTrait;
 
 /**
  * Tests taxonomy terms remove from publication.
@@ -12,6 +13,8 @@ use Drupal\bibcite_entity\Entity\Reference;
  * @covers \Drupal\cp_taxonomy\Form\RemoveTermsFromPublicationForm
  */
 class RemoveTermsPublicationTest extends CpTaxonomyExistingSiteJavascriptTestBase {
+
+  use CpTaxonomyTestTrait;
 
   protected $term;
   protected $groupAdmin;
@@ -52,7 +55,7 @@ class RemoveTermsPublicationTest extends CpTaxonomyExistingSiteJavascriptTestBas
     $page = $this->getCurrentPage();
     $page->findField('bibcite_reference_bulk_form[0]')->check();
     $page->findField('bibcite_reference_bulk_form[1]')->check();
-    $this->removeTermWithAction();
+    $this->removeTermWithAction('cp_taxonomy_remove_terms_bibcite_reference_action');
     $web_assert->pageTextContains('No term was removed from the content');
     $web_assert->pageTextContains('Taxonomy term ' . $this->term->label() . ' was removed from the content');
     $warning_wrapper = $page->find('css', '.messages--warning');
@@ -65,31 +68,6 @@ class RemoveTermsPublicationTest extends CpTaxonomyExistingSiteJavascriptTestBas
     $saved_publication = Reference::load($publication1->id());
     $term_value = $saved_publication->get('field_taxonomy_terms')->getString();
     $this->assertEmpty($term_value);
-  }
-
-  /**
-   * Helper function will select a term and remove from selected publications.
-   */
-  protected function removeTermWithAction() {
-    $web_assert = $this->assertSession();
-    $page = $this->getCurrentPage();
-    $select = $page->findField('action');
-    $select->setValue('cp_taxonomy_remove_terms_bibcite_reference_action');
-    $page->pressButton('Apply to selected items');
-    $web_assert->statusCodeEquals(200);
-    $page = $this->getCurrentPage();
-    $select = $page->findField('vocabulary');
-    $select->setValue('vocab_group_1');
-    $this->waitForAjaxToFinish();
-    $page->find('css', '.chosen-search-input')->click();
-
-    $result = $web_assert->waitForElementVisible('css', '.active-result.highlighted');
-    $this->assertNotEmpty($result, 'Chosen popup is not visible.');
-    $web_assert->pageTextContains($this->term->label());
-    $page->find('css', '.active-result.highlighted')->click();
-    $page->find('css', '.chosen-search-input')->click();
-    $page->pressButton('Remove');
-    $web_assert->statusCodeEquals(200);
   }
 
 }
